[raw file config] Add LIST_ITEM_IDENTIFIER_TAG to unique identify items in raw file config attribute's list (Recidiviz/recidiviz-data#45353)
## Description of the change

firstly, move raw file config enums to a separate enum file and move
`is_meaningful_docstring` to utils file because the
`raw_file_configs.py` file is getting quite large

secondly, in the new raw data config writer we are converting raw file
config objects to attribute dictionaries and then determining the
differences between two config objects by comparing the dictionaries.
this can be tricky for attributes that are `list[child_class]` because
we only want to update the leaf attributes that have changed and not the
entire list. therefore, we need a way to meaningfully identify an item
in a list from version1 to version2 since lists are not guaranteed to
maintain ordering with the new items at the end (ex we alphabetize known
values)

for example say we have (ignore the random accidental double spacing in
these examples)
```
    known_values:
      - value: AB
        description: |-
                Alberta
      - value: AL
        description:  |-
                Alabama
      - value: AM
        description:  |-
                American Samoa
      - value: AR
        description:  |-
                Arkansas
      - value: AZ
        description:  |-
                Arizona
.....
```
and we updated AB -> Albania and added AK, we would want the output to
be
```
    known_values:
      - value: AB
        description: Albania
      - value: AK
        description: Alaska
      - value: AL
        description:  |-
                Alabama
      - value: AM
        description:  |-
                American Samoa
      - value: AR
        description:  |-
                Arkansas
      - value: AZ
        description:  |-
                Arizona
....
```
if we updated the known_values list like `known_values =
new_known_values` we would end up with reformats like
```
    known_values:
      - value: AB
        description: Albania
      - value: AK
        description: Alaska
      - value: AL
        description:  Alabama
      - value: AM
        description:  American Samoa
      - value: AR
        description:  Arkansas
      - value: AZ
        description: Arizona
```
which can get really noisy and so we need a way to distinguish between
elements of a list that were updated/deleted/added or else we would end
up with
```
    known_values:
      - value: AL
        description:  |-
                Alabama
      - value: AM
        description:  |-
                American Samoa
      - value: AR
        description:  |-
                Arkansas
      - value: AZ
        description:  |-
                Arizona
      - value: AB
        description: Albania
      - value: AK
        description: Alaska
```

instead, add `LIST_ITEM_IDENTIFIER_TAG` metadata tag to attributes than
be used to uniquely identify an object in a list. then add a post init
validation that recursively checks that all list-typed fields in
`DirectIngestRawFileConfig` and its child classes refer to element
classes that contain exactly one field marked with metadata
`{LIST_ITEM_IDENTIFIER_TAG: True}`.

there is an oversight here in that the validation assumes that there are
no child class references other than direct references and those found
in lists (ex ignores child classes in dicts, tuples, etc), but doesn't
enforce this to be true, but it is currently true and i can add that
validation in a future PR.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#45352

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: ab84bf31a3b57996e81c4743e622adee737dfd85