[US_MI] Add 2 month substance free milestone (Recidiviz/recidiviz-data#44242)
## Updates 07/13/2025
* There used to be 29689 people eligible for this -- now there are
14004. This is because I fixed a big where streaks weren't ending after
1 year, even if there were no new negative drug tests.
* I think I fixed the TES criteria! Before, the TES criteria tracked if
a person was currently on a negative streak, so it could correctly
report that as of 2025-07-07, you had only had negative drug tests.
However, it wasn't accurately reflecting when the most recent drug test
was -- now it can say you are on a negative streak as of 2025-07-07, and
your last negative test was 2025-01-01. If you're tested again on
2025-07-08, we'll see a new TES record which says you are still on a
streak, but now your most recent test was on 2025-07-08.

## Description of the change
Adds a 2 month substance free milestone for Michigan. It achieves this
by creating a drug_screens_preprocessed view, which feeds a TES
criteria, which feeds into ClientRecord.

### Notes
1. The preprocessed view does some deduping. This basically copies the
logic used in IX and TN. It dedupes multiple drug test results of the
same type and on the same date by prioritizing positive results. So if
you have 3 breathalyzers on the same day, and one comes back positive,
we say that you had a positive drug test that day.
~**Exception:** The exception to above is that you can have a positive
on-site UA (coded 3710), and a negative lab UA (coded 3711) on the same
day, and in this case, we will prioritize the lab UA since it seems like
lab UA would be more accurate than on-site -- so negative lab UA trumps
positive on-site UA. For any other non-urine tests, we prioritize
positive results over negative results if there are multiple tests on a
given day.~
2. Someone meets this milestone if:
A) They have only had negative drug tests for at least 2 months. Any
positive ends the streak.
  B) They have had a drug test within at least the last year.
This is implemented as "streaks" -- you are always on a streak of
positive or negative drug tests, and streaks expire after a year of
inactivity. If you are on a negative streak for at least 2 months, you
are highlighted in this milestone.

**Examples:**
1. If you had only negative drug tests, but your last test was in 1998,
we don't highlight you with this milestone because it's been over a year
since your last negative drug test.
2. If your last drug test was negative, and it happened exactly one year
ago, you'd be in this milestone today and out tomorrow. If you then got
another negative drug test a month later, we would wait two months from
that test date before highlighting you for this milestone, because your
previous streak has expired and it takes 2 months for a streak to be
considered eligible for this milestone.

### How the TES algorithm works
1. We create periods which reflect the status of the most recent test --
IE from 2025-07-01 to 2025-07-09 you tested positive, then from
2025-07-09 until 2025-07-15 you tested negative, then from 2025-07-29
you tested negative. Each test creates a new period.
2. We use aggregate_adjacent_spans to create spans during which someone
was on a negative or positive drug test streak.
3. We join these streaks back into the periods created in step one to
join in the streak start date. There's some complexity here, because we
may actually have to create a new period record because we offset when
the streak begins. So if you tested negative on 2025-07-01, then tested
negative again on 2025-08-01, we want to make sure your streak begins on
2025-09-01 (2 months after first negative test), but also reflect that
the most recent test was 2025-08-01. I'm not confident the algorithm for
accomplishing that nuance is as simple as possible, but I tried to
clearly lay out that this is what is happening in the comments!

**Tests**
It's mostly eye-balling for testing changes like this. Things that make
me feel like this is okay:
1. Nobody is added or removed from client record
2. There are clients in client record in MI -- 14004 meet this milestone
(if this feels high, we could adjust how soon streaks expire)
3. [I picked a few random
people](https://console.cloud.google.com/bigquery?ws=!1m7!1m6!12m5!1m3!1srecidiviz-staging!2sus-central1!3s2a6d9727-abed-4e95-ae24-781f181fa57b!2e1),
and their milestones appeared correct when compared with raw data
:shrug:
4. Loaded new client_record into : zaaron_mi_views.

### Recidiviz-Dashboard PR
This full milestones functionality also depends on [this
recidiviz-dashboards
PR](https://github.com/Recidiviz/recidiviz-dashboards/pull/8782), but
this (the PR you are reading) in recidiviz-data PR should be merged
first.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#43364

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 838f3a61bf348745e1430a3cc6845c2ea64be795