[US_TN] Hydrate current_state_provided_start_date (SentenceEffectiveDate) (Recidiviz/recidiviz-data#45387)
## Description of the change

* Hydrates `current_state_provided_start_date` via the _most recent_
`SentenceEffectiveDate` from TN
* Flips the switch in normalization to infer net-new SERVING statuses
for these dates, if needed. Note I haven't updated status ingest yet, so
there are no pre-normalization statuses from effective date
* Handles the edgecase where the state tells us a sentence starts and
stops on the same day. We create the net new serving status in a way
that it must be before the completion status (but on the same day). I
add a test for this test case because I would like to see how dates and
credits turn out for this.
* I turned off auto-generated IDs in favor of the explicit setting of
the ID in normalization only. I found this way easier to debug.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Supports
[Recidiviz/recidiviz-data#44483](https://github.com/Recidiviz/recidiviz-data/issues/44483)

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8b113396321a4346b45b9e1b7a7518b0096f16fc