[Resource Search] Script to Scrape Idaho Websites (Recidiviz/recidiviz-data#45461)
This PR introduces the core script responsible for scraping resource
data and populating our Cloud SQL PostgreSQL database. This step in
establishes the primary source of truth for our resource data, which
will subsequently be transferred to BigQuery.

The cool thing is, the script itself is actually pretty simple. It's
just calling that Python scrape script I've already got working great
locally. The main difference now is that the data it collects will be
saved right into the Cloud SQL instance. We're able to keep the new code
minimal because we're really leaning on the existing
`recidiviz/resource-search/src/db.py` logic.
```
async def initialize_global_engine() -> None:
    """Initializes the global AsyncEngine once."""
    global _global_engine
    if _global_engine is None:
        db_url_to_use = (
            local_postgres_helpers.async_on_disk_postgres_db_url(
                username=get_secret_from_local_directory(
                    secret_id="resource_search_db_user"  # nosec
                ),
                port=get_secret_from_local_directory(
                    secret_id="resource_search_db_port"  # nosec
                ),
                password=get_secret_from_local_directory(
                    secret_id="resource_search_db_password"  # nosec
                ),
                database=get_secret_from_local_directory(
                    secret_id="resource_search_db_host"  # nosec
                ),
            )
            if not in_gcp_staging() or not in_gcp_production()
            else SQLAlchemyAsyncEngineManager.get_server_postgres_asyncpg_instance_url(
                database_key=SQLAlchemyDatabaseKey.for_schema(
                    SchemaType.RESOURCE_SEARCH
                ),
            )
        )
        _global_engine = (
            await SQLAlchemyAsyncEngineManager.init_async_engine_for_postgres_instance(
                db_url=db_url_to_use,
                database_key=SQLAlchemyDatabaseKey.for_schema(
                    SchemaType.RESOURCE_SEARCH
                ),
            )
        )
 ```

Just a heads-up on testing: since this involves live connections to Cloud SQL in GCP, I haven't been able to do a full end-to-end test locally. I'll get that comprehensive testing done once this PR is merged and deployed to our staging environment. That's where we'll confirm everything is flowing as it should, from scraping all the way to data saving in Cloud SQL.

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally Identifiable Information (PII) being mistakenly added in this pull request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful to a reviewer
- [ ] Potential security implications or infrastructural changes have been considered, if relevant

---------

Co-authored-by: Helper Bot <helperbot@recidiviz.org>
GitOrigin-RevId: f4d52f0efc40a2a10d2c5a7c78f40fefaacab546