[US_UT] Raw data configs for new CAP-related data (Recidiviz/recidiviz-data#45555)
## Description of the change

This adds raw data configuration files for these five new raw data
tables:
- ofndr_cap_goal_actn_stp
- cap_srvce_priority
- cap_actn_stp_trgt
- cap_earned_time_credit_balance
- cap_cstm_actn_stp

I successfully uploaded each of these files to a sandbox with the prefix
`elise`.

This PR also makes a minor change to the usage example in the
`import_raw_files_to_sandbox` docstring to put quotation marks instead
of parentheses when multiple file tags are included in the
`--file-tag-filter-regex` arg.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 75106fb058f7bded03e5b1a89964329890b3c92e