[looker] Call `create_looker_pr_if_not_exists` if looker remote branch exists (Recidiviz/recidiviz-data#45542)
## Description of the change

before, the `refresh_recidiviz_data_to_looker_sync_pr` script would exit
if there were no looker changes to be committed to the looker repo
branch.

however, if a new commit is pushed while the workflow is still running
it will cancel the workflow. this might leaves us in a state where the
looker changes have been pushed to a branch but the looker PR has not
yet been opened. then, the subsequent run might not detect any changes
and so it would just exit.

this PR updates the script to call `create_looker_pr_if_not_exists` and
`upsert_recidiviz_data_pr_comment` if the looker remote branch exists,
regardless of whether there were new changes to commit on this workflow
run in order to ensure a PR always gets opened

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 54dece6bd3b4b94b069c04f677f8900e62dc95cd