[CI] Only use `tmpfs` extra for the BigQuery emulator tests (Recidiviz/recidiviz-data#45618)
## Description of the change

The [`tmpfs` extra](https://runs-on.com/caching/tmpfs/) (also known as
YOLO mode) mounts a volume to the instance which allows disk
writes/reads to happen in-memory. This means that if one of the
directories consumes more memory than is available to the instance the
job will hang.

It yielded great improvements to the BigQuery Emulator jobs but causes
flakiness in some of our other jobs that are not assigned as many
resources (see Recidiviz/recidiviz-data#45518)

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: a5c8902398eb4c813c84dff0b207bdc0721951b2