[US_AZ] Check TABE results in mandatory literacy criteria (Recidiviz/recidiviz-data#45519)
## Description of the change

This adds a source to the `meets_mandatory_literacy` query fragment used
to populate our `US_AZ_MEETS_FUNCTIONAL_LITERACY_TPR` and
`US_AZ_MEETS_FUNCTIONAL_LITERACY_DTP` criteria queries. The addition
checks `state_assessment` for a TABE test in which the person scored at
an 8th grade or above level on each section, which is established
[here](https://corrections.az.gov/office-deputy-director/education-programs-and-community-reentry/education?utm_source=chatgpt.com)
as the minimum qualification for being exempt from the functional
literacy course.

The logic for finding the date on which to start counting them as having
passed the TABE is a little complex. The TLDR is that a person can pass
each section on different days, and they are considered to have "passed
TABE" as soon as they pass all three sections, either at once or
separately. We use the earliest date on which they passed all three
sections to count them as meeting this criteria.

This also adds a `StateAssessmentType` for TABE tests so they are easier
to identify.

I loaded this change into a sandbox and it makes one person newly
eligible and one person newly almost eligible.

<img width="994" height="300" alt="Screenshot 2025-07-23 at 3 37 39 PM"
src="https://github.com/user-attachments/assets/04993f21-a178-41d7-b4f2-5a751692c56b"
/>

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: df26caec0ecf7b508f6c3874545a4c739e9b203a