[raw data config writer] Implement raw data ruamel yaml update (Recidiviz/recidiviz-data#45539)
## Description of the change

This PR adds a function and its helpers that help update yaml
dictionaries for raw data configurations. This is done by comparing an
updated dictionary against the original dictionary and only applying the
identified differences to the yaml dictionary.

Adds basic unit tests but will add more comprehensive test suite once
top level `update_region_raw_file_yamls` function is implemented.

Leaving list case as a TODO since it is more complex to compare two
lists of dictionaries that are not guaranteed to be the same length/in
the same order.

This case will be implemented once more comprehensive UTs are written.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#44663

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Carol Chen <carolchen@Carols-MacBook-Pro.local>
GitOrigin-RevId: 7f59641afb1f4dfe1d9d9a97e489f5c40a831426